Revert "fib_rules: return 0 directly if an exactly same rule exists when NLM_F_EXCL not supplied"
This reverts commit 1fff19a925e524556e85efcd728efad5274ce5b6.

Some users are having soft reboots with loss of connectivity because
netd does not like this change. I cannot reproduce this on my OnePlus 6
with OxygenOS and all of the reports I have gotten so far appear to be
from custom ROM developers so maybe it is a commit that they have added
in userspace causing issues. I would report this upstream for
verification but Android is always special and I don't currently have
the time to help debug or get a fix tested (especially if I can't
reproduce on my current setup).

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: khusika <khusikadhamar@gmail.com>